Retrieve and update local and/or remote data from formr

This function generates a data frame with participant's responses to each item, along with some session-specific metadata. It takes participants (the output of bvq_participants()) and runs (a character vector that can take zero, one, or multiple of the following values: "formr2", "formr-short", "formr-lockdown") as arguments.

Usage

bvq_responses_vocab(participants = bvq_participants())

Arguments

participants

Participants data frame, as generated by bvq_participants(). If NULL (default), bvq_participants() is run.

Value

A data frame (actually, a tibble::tibble containing participant's responses to each item, along with some session-specific metadata. The output includes the following variables:

child_id: a character string with five digits indicating a participant's identifier in the database from the Laboratori de Recerca en Infància at Universitat Pompeu Fabra. This value is always the same for each participant, so that different responses from the same participant share the same child_id.

response_id: a character string identifying a single response to the questionnaire. This value is always unique for each response to the questionnaire, even for responses from the same participant.

time: a numeric value indicating how many times a given participant has been sent the questionnaire, regardless of whether they completed it or not.

version: a character string indicating what version of the questionnaire a given participant filled in. Different versions may contain a different subset of items, and the administration instructions might vary slightly (see formr questionnaire templates in the GitHub repository). Also, different versions were designed, implemented, and administrated at different time points (e.g., before/during/after the COVID-related lockdown).

version_list: a character string indicating the specific list of items a participant was assigned to. Only applies in the case of short versions of BVQ, such as bvq-short, bvq-long, bvq-lockdown, or bvq-1.0.0, where the list of items was partitioned into several versions.#' * item: character string indicating the item identifier (e.g., spa_mesa). This value is unique for each item. Responses to the same item from different participants are linked by the same item value.

response: integer indicating the participant's response to a give item: 1 if "No" (the participant does not understand or produce the word), 2 if "Understands" (the participants understands the word), or 3 if "Understands and Says" (the participant understands and produces the item).

date_birth: lubridate::Date indicating participants birth date.

date_started: lubridate::Date indicating when participants logged to the questionnaire for the first time.

date_finished: lubridate::Date indicating when participants logged to the questionnaire for the last time.

sex: a character string indicating participants' biological sex, as reported by the parents.

doe_spanish: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Spanish, as estimated by parents/caregivers This value aggregates participants' exposure to any variant of Spanish (e.g., European and American Spanish).

doe_catalan: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Catalan, as estimated by parents/caregivers This value aggregates participants' exposure to any variant of Catalan (e.g., Catalan from Majorca or Barcelona).

doe_others: a numeric value ranging from 0 to 1 indicating participants' daily exposure to languages other than Spanish or Catalan, as estimated by parents/caretakers, aggregating participants' exposure to all those other languages (e.g., Norwegian, Arab, Swahili).

edu_parent1: a character string indicating the educational attainment of one of the parents/caretakers.

edu_parent2: a character string indicating the educational attainment of the other parent/caretaker, if any.

Author

Gonzalo Garcia-Castro

Examples

if (FALSE) { # \dontrun{
bvq_responses_vocab()
} # }