Generate item-level norms for age, sex, language profile and item dominance

This function generates a data frame with the estimated proportion of children that understand and/or produce some items for a selected age range and participant profiles. Estimated proportions and corresponding standard errors and confidence intervals are computed adjusting for zero- and one-inflation (see function prop_adj()).

Usage

bvq_norms(
  participants,
  responses,
  item = NULL,
  language = c("Catalan", "Spanish"),
  type = c("understands", "produces"),
  age = NULL,
  lp = c("Bilingual", "Monolingual", "Other"),
  sex = c("Female", "Male"),
  semantic_category = NULL,
  .width = 0.95,
  ...
)

Arguments

participants

Participants data frame, as generated by bvq_participants().

responses

Responses data frame, as generated by bvq_responses().

item

Character string indicating the item to compute norms for. If left NULL (default) norms will be computed for all items. You can check the available items in the pool data set running data("pool").

language

Character string indicating the language to compute vocabulary norms for: "catalan" and/or "spanish".

type

Character string indicating the vocabulary type to compute norms for. Takes "understands" and/or "produces" (defaults to both).

age

Numeric vector of length two (min-max) indicating the age range of participants to compute norms for.

lp

character string indicating the language profile of participants to compute norms for: "Bilingual", "Monolingual", "Other" (defaults to all).

sex

character string indicating the sex of participants to compute norms for. Takes "Female" and/or "Male" (defaults to both).

semantic_category

character string indicating the semantic/functional category or categories to include items from. See available categories in the pool data set by running data("pool").

.width

Numeric values ranging from 0 to 1 (not included) indicating the confidence level of confidence intervals (defaults to 0.95).

...

Unused

Value

A data frame (actually, a tibble::tibble with the proportion of participants in the sample that understand or produce the items indicated in item, along with the standard error and confidence interval of the estimation. The output contains the following variables:

te: an integer identifying the translation equivalent (a.k.a., pair of cross-language synonyms, doublets) the item belongs to.

item: character string indicating the item identifier (e.g., spa_mesa). This value is unique for each item. Responses to the same item from different participants are linked by the same item value.

language: a character string indicating the language the item response belongs to: "Catalan" if item in Catalan), "Spanish" if item in Spanish.

age: an integer indicating the age of participants (in months) for which the estimates should be computed. If a non-integer is provided (e.g., 15.36, it is rounded downwards using floor().)

type: a character string indicating the vocabulary type computed: "understands" if option 'Understands' was selected, and "produces" if option 'Understands & Says' was selected.

lp: a character string indicating participants' language profile, classified using parental reports of language exposure (see doe_spanish, "doe_catalan, and doe_others), and the thresholds passed in the bilingual_threshold and other_threshold.

semantic_category: a character string indicating the semantic/function category the item belongs to (e.g., "Vehicles", "Actions").

item_dominance: a character string that takes the value "L1" if the item belongs to participants' language of most exposure, and L2 if the item belongs to participants' language of least exposure.

label: a character string indicating the text presented to participants in the questionnaire (replacing the item identifier).

.sum: a positive integer indicating the number of positive responses: responses is 2 (Understands) or 3 (Understands & Says) for type = 'understands', and 3 (Understands & Says) if type = 'produces'.

.n: a positive integer indicating the total number number of responses (useful for computing proportions).

.prop: a numeric value ranging from 0 to 1 (both included) indicating the estimated proportion of participants that provided a positive response, adjusted following Gelman et al.'s method to account for zero- and one-inflation (see function prop_adj).

Author

Gonzalo Garcia-Castro